Move predicate segment util to context extractor
