+++ infra/binder/src/main/java/org/apache/shardingsphere/infra/binder/context/extractor/SQLStatementContextExtractor.java
@@ -20,18 +20,28 @@ package org.apache.shardingsphere.infra.binder.context.extractor;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.infra.annotation.HighFrequencyInvocation;
+import org.apache.shardingsphere.infra.binder.context.segment.insert.values.InsertSelectContext;
 import org.apache.shardingsphere.infra.binder.context.statement.SQLStatementContext;
+import org.apache.shardingsphere.infra.binder.context.statement.dml.InsertStatementContext;
+import org.apache.shardingsphere.infra.binder.context.statement.dml.SelectStatementContext;
 import org.apache.shardingsphere.infra.binder.context.type.IndexAvailable;
 import org.apache.shardingsphere.infra.binder.context.type.TableAvailable;
+import org.apache.shardingsphere.infra.binder.context.type.WhereAvailable;
 import org.apache.shardingsphere.infra.database.core.type.DatabaseType;
 import org.apache.shardingsphere.infra.metadata.database.ShardingSphereDatabase;
 import org.apache.shardingsphere.infra.metadata.database.schema.QualifiedTable;
 import org.apache.shardingsphere.infra.metadata.database.schema.util.IndexMetaDataUtils;
 import org.apache.shardingsphere.sql.parser.statement.core.segment.ddl.index.IndexSegment;
+import org.apache.shardingsphere.sql.parser.statement.core.segment.dml.column.ColumnSegment;
+import org.apache.shardingsphere.sql.parser.statement.core.segment.dml.expr.BinaryOperationExpression;
+import org.apache.shardingsphere.sql.parser.statement.core.segment.dml.predicate.WhereSegment;
 
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.LinkedHashMap;
 import java.util.LinkedList;
+import java.util.Map;
 
 @HighFrequencyInvocation
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
@@ -59,6 +69,81 @@ public final class SQLStatementContextExtractor {
         for (QualifiedTable each : IndexMetaDataUtils.getTableNames(database, databaseType, indexes)) {
             result.add(each.getTableName());
         }
+        return result;
+    }
+    
+    /**
+     * Get all subquery contexts.
+     *
+     * @param sqlStatementContext SQL statement context
+     * @return all subquery contexts
+     */
+    public static Collection<SelectStatementContext> getAllSubqueryContexts(final SQLStatementContext sqlStatementContext) {
+        Collection<SelectStatementContext> result = new LinkedList<>();
+        if (sqlStatementContext instanceof SelectStatementContext) {
+            result.addAll(((SelectStatementContext) sqlStatementContext).getSubqueryContexts().values());
+            ((SelectStatementContext) sqlStatementContext).getSubqueryContexts().values().forEach(each -> result.addAll(getAllSubqueryContexts(each)));
+        }
+        if (sqlStatementContext instanceof InsertStatementContext && null != ((InsertStatementContext) sqlStatementContext).getInsertSelectContext()) {
+            InsertSelectContext insertSelectContext = ((InsertStatementContext) sqlStatementContext).getInsertSelectContext();
+            result.add(insertSelectContext.getSelectStatementContext());
+            result.addAll(insertSelectContext.getSelectStatementContext().getSubqueryContexts().values());
+            insertSelectContext.getSelectStatementContext().getSubqueryContexts().values().forEach(each -> result.addAll(getAllSubqueryContexts(each)));
+        }
+        return result;
+    }
+    
+    /**
+     * Get all where segments.
+     *
+     * @param sqlStatementContext SQL statement context
+     * @return all where segments
+     */
+    public static Collection<WhereSegment> getAllWhereSegments(final SQLStatementContext sqlStatementContext) {
+        if (!(sqlStatementContext instanceof WhereAvailable)) {
+            return Collections.emptySet();
+        }
+        Collection<SelectStatementContext> allSubqueryContexts = getAllSubqueryContexts(sqlStatementContext);
+        return SQLStatementContextExtractor.getWhereSegments((WhereAvailable) sqlStatementContext, allSubqueryContexts);
+    }
+    
+    /**
+     * Get all where segments.
+     *
+     * @param whereAvailable where available
+     * @param allSubqueryContexts all subquery contexts
+     * @return all where segments
+     */
+    public static Collection<WhereSegment> getWhereSegments(final WhereAvailable whereAvailable, final Collection<SelectStatementContext> allSubqueryContexts) {
+        Map<Integer, WhereSegment> uniqueWhereSegments = new LinkedHashMap<>();
+        whereAvailable.getWhereSegments().forEach(each -> uniqueWhereSegments.put(each.getStartIndex() + each.getStopIndex(), each));
+        allSubqueryContexts.forEach(each -> each.getWhereSegments().forEach(where -> uniqueWhereSegments.put(where.getStartIndex() + where.getStopIndex(), where)));
+        return new ArrayList<>(uniqueWhereSegments.values());
+    }
+    
+    /**
+     * Get all column segments.
+     *
+     * @param whereAvailable where available
+     * @param allSubqueryContexts all subquery contexts
+     * @return all column segments
+     */
+    public static Collection<ColumnSegment> getColumnSegments(final WhereAvailable whereAvailable, final Collection<SelectStatementContext> allSubqueryContexts) {
+        Collection<ColumnSegment> result = new LinkedList<>(whereAvailable.getColumnSegments());
+        allSubqueryContexts.forEach(each -> result.addAll(each.getColumnSegments()));
+        return result;
+    }
+    
+    /**
+     * Get all join conditions.
+     *
+     * @param whereAvailable where available
+     * @param allSubqueryContexts all subquery contexts
+     * @return all join conditions
+     */
+    public static Collection<BinaryOperationExpression> getJoinConditions(final WhereAvailable whereAvailable, final Collection<SelectStatementContext> allSubqueryContexts) {
+        Collection<BinaryOperationExpression> result = new LinkedList<>(whereAvailable.getJoinConditions());
+        allSubqueryContexts.forEach(each -> result.addAll(each.getJoinConditions()));
         return result;
     }
 }
